fix: update the stories & docs links
