+++ modules/react/loading-dots/stories/LoadingDots.mdx
@@ -1,9 +1,9 @@
-import {ExampleCodeBlock} from '@workday/canvas-kit-docs';
+import {ExampleCodeBlock, SymbolDoc} from '@workday/canvas-kit-docs';
 import {Basic} from './examples/Basic';
 import {RTL} from './examples/RTL';
 import {Accessible} from './examples/Accessible';
-import {DarkBackgrounds} from './examples/DarkBackgrounds';
 import {CustomShape} from './examples/CustomShape';
+import {CustomColorAndAnimation} from './examples/CustomColorAndAnimation';
 import * as LoadingDotsStories from './LoadingDots.stories';
 
 <Meta of={LoadingDotsStories} />
@@ -13,7 +13,7 @@ import * as LoadingDotsStories from './LoadingDots.stories';
 Loading Dots make users aware that content is currently being loaded, processing, or that change
 will occur on the page.
 
-[> Workday Design Reference](https://design.workday.com/components/indicators/loading-dots)
+[> Workday Design Reference](https://canvas.workdaydesign.com/components/indicators/loading-dots/)
 
 ## Installation
 
@@ -31,18 +31,18 @@ yarn add @workday/canvas-kit-react
 
 <ExampleCodeBlock code={RTL} />
 
-### Dark Backgrounds
-
-<ExampleCodeBlock code={DarkBackgrounds} />
-
 ### Custom Shape
 
 <ExampleCodeBlock code={CustomShape} />
 
+### Custom Color and Animation
+
+<ExampleCodeBlock code={CustomColorAndAnimation} />
+
 ### Custom Styles
 
 Loading Dots supports custom styling via the `cs` prop. For more information, check our
-["How To Customize Styles"](https://workday.github.io/canvas-kit/?path=/docs/styling-how-to-customize-styles--page).
+["How To Customize Styles"](https://workday.github.io/canvas-kit/?path=/docs/styling-how-to-customize-styles--docs).
 
 Custom styling is also supported through the [Loading Dots documented props below](#props).
 
@@ -62,11 +62,6 @@ components included in Canvas to describe both the appearance and disappearance 
 
 <ExampleCodeBlock code={Accessible} />
 
-## Props
+## Component API
 
-Loading Dots exposes the following props:
-
-- `loadingDotsColors`: Applies backgroundColor to loading dots, intended for use with the circle variant design on grey/dark/image-based backgrounds. Defaults to `system.color.bg.alt.strong`.
-- `animationDurationMs`: Duration of the loading animation. Defaults to `40ms`.
-
-Other undocumented props are spread to its outermost element.
+<SymbolDoc name="LoadingDots" fileName="/react/" />
