+++ modules/react/loading-dots/stories/LoadingDots.stories.ts
@@ -7,8 +7,8 @@ import {LoadingDots} from '@workday/canvas-kit-react/loading-dots';
 import {Basic as BasicExample} from './examples/Basic';
 import {RTL as RTLExample} from './examples/RTL';
 import {Accessible as AccessibleExample} from './examples/Accessible';
-import {DarkBackgrounds as DarkBackgroundsExample} from './examples/DarkBackgrounds';
 import {CustomShape as CustomShapeExample} from './examples/CustomShape';
+import {CustomColorAndAnimation as CustomColorAndAnimationExample} from './examples/CustomColorAndAnimation';
 
 export default {
   title: 'Components/Indicators/Loading Dots',
@@ -32,9 +32,9 @@ export const RTL: Story = {
 export const Accessible: Story = {
   render: AccessibleExample,
 };
-export const DarkBackgrounds: Story = {
-  render: DarkBackgroundsExample,
-};
 export const CustomShape: Story = {
   render: CustomShapeExample,
+};
+export const CustomColorAndAnimation: Story = {
+  render: CustomColorAndAnimationExample,
 };
