+++ modules/react/loading-dots/stories/examples/CustomColorAndAnimation.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import {LoadingDots} from '@workday/canvas-kit-react/loading-dots';
+import {system} from '@workday/canvas-tokens-web';
+import {createStyles} from '@workday/canvas-kit-styling';
+
+const styleOverrides = {
+  parentContainer: createStyles({
+    display: 'flex',
+    gap: system.space.x4,
+  }),
+};
+
+export const CustomColorAndAnimation = () => {
+  return (
+    <div className={styleOverrides.parentContainer}>
+      <LoadingDots loadingDotColor={system.color.fg.primary.default} animationDurationMs="60ms" />
+    </div>
+  );
+};
