--- modules/react/loading-dots/stories/examples/DarkBackgrounds.tsx
@@ -1,33 +1,0 @@
-import React from 'react';
-import {LoadingDots, loadingDotsStencil} from '@workday/canvas-kit-react/loading-dots';
-import {system} from '@workday/canvas-tokens-web';
-import {createStyles, createStencil} from '@workday/canvas-kit-styling';
-
-const styleOverrides = {
-  parentContainer: createStyles({
-    display: 'flex',
-    gap: system.space.x4,
-  }),
-};
-
-const loadingStencil = createStencil({
-  base: {
-    borderRadius: system.shape.round,
-    backgroundColor: system.color.bg.overlay,
-    height: 80,
-    width: 80,
-    alignItems: 'center',
-    justifyContent: 'center',
-    display: 'flex',
-    [loadingDotsStencil.vars.loadingDotColor]: system.color.bg.alt.strong,
-    [loadingDotsStencil.vars.animationDurationMs]: '40ms',
-  },
-});
-
-export const DarkBackgrounds = () => {
-  return (
-    <div className={styleOverrides.parentContainer}>
-      <LoadingDots cs={loadingStencil()} />
-    </div>
-  );
-};
